requirements.txt
flask
torch
diffusers
Pillow
# Add other dependencies as needed
